refactor: implemented some other changes to the publishing system
